+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/member/dto/request/MemberUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.member.dto.request;
+package com.mobble.mobbleserver.infrastructure.web.member.dto.request;
 
 public record MemberUpdateRequestDto(Long groundCode, String profileImage) {
 }
